separa o app e o server
